requirements.txt
asgiref==3.7.2
certifi==2022.6.15
cffi==1.15.1
chardet==5.2.0
charset-normalizer==2.1.0
cryptography==37.0.4
defusedxml==0.7.1
Django==4.2.5
django-allauth==0.51.0
django-appconf==1.0.5
django-autoslug==1.9.8
django-cors-headers==3.13.0
django-environ==0.9.0
django-etc==1.3.1
django-hitcount==1.3.5
django-imagekit==4.1.0
django-jazzmin==2.5.0
django-materializecss-form==1.1.17
django-qrcode==0.3
django-ranged-response==0.2.0
django-simple-captcha==0.6.1
django-simple-history==3.8.0
django-unfold==0.57.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
idna==3.3
oauthlib==3.2.0
pilkit==3.0
Pillow==10.0.1
psycopg2-binary==2.9.10
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
qrcode==8.0
reportlab==4.2.5
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.7.1
urllib3==1.26.10
python-dateutil
